Refactored and fixed methods, major changes to update method.

+++ src/main/java/com/example/demo/service/DefaultCarPartService.java
@@ -1,12 +1,18 @@
 package com.example.demo.service;
 
+import com.example.demo.entity.CarModel;
 import com.example.demo.entity.CarPart;
+import com.example.demo.exception.CarModelNotFoundException;
+import com.example.demo.exception.CarPartNotFoundException;
 import com.example.demo.repository.CarPartRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
 import java.util.Objects;
+import java.util.Optional;
 
 @Service
 public class DefaultCarPartService implements CarPartService {
@@ -20,8 +26,17 @@ public class DefaultCarPartService implements CarPartService {
     }
 
     @Override
-    public CarPart getCarPartById(Long id) {
-        return carPartRepository.findById(id).orElse(null);
+    public Page<CarPart> getAllCarParts(Pageable pageable) {
+        return carPartRepository.findAll(pageable);
+    }
+
+    @Override
+    public Optional<CarPart> getCarPartById(Long id) {
+        Optional<CarPart> carPart = carPartRepository.findById(id);
+        if (carPart.isEmpty()) {
+            throw new CarPartNotFoundException();
+        }
+        return carPart;
     }
 
     @Override
@@ -30,32 +45,43 @@ public class DefaultCarPartService implements CarPartService {
     }
 
     @Override
-    public CarPart updateCarPart(CarPart carPart, Long id) {
-        CarPart carPartDB = carPartRepository.findById(id).get();
-
-        if (Objects.nonNull(carPartDB.getName()) && !carPartDB.getName().isEmpty()) {
-            carPartDB.setName(carPartDB.getName());
+    public void updateCarPart(CarPart carPart, Long id) {
+        Optional<CarPart> optionalCarPart = carPartRepository.findById(id);
+        if (optionalCarPart.isPresent()) {
+            CarPart carPartToUpdate = optionalCarPart.get();
+            if (Objects.nonNull(carPart.getName()) && !carPart.getName().isEmpty()) {
+                carPartToUpdate.setName(carPart.getName());
+            }
+            if (carPart.getPrice() >= 0) {
+                carPartToUpdate.setPrice(carPart.getPrice());
+            }
+            if (carPart.getQuantity() >= 0) {
+                carPartToUpdate.setQuantity(carPart.getQuantity());
+            }
+            carPartRepository.updateCarPart(carPartToUpdate.getName(), carPartToUpdate.getPrice(), carPartToUpdate.getQuantity(), id);
+        } else {
+            throw new CarPartNotFoundException();
         }
-
-        if (Objects.nonNull(carPartDB.getPrice())) {
-            carPartDB.setPrice(carPartDB.getPrice());
-        }
-
-        if (Objects.nonNull(carPartDB.getQuantity())) {
-            carPartDB.setQuantity(carPartDB.getQuantity());
-        }
-
-        return carPartRepository.save(carPartDB);
     }
 
     @Override
     public void deleteCarPart(Long id) {
-        carPartRepository.deleteById(id);
+        Optional<CarPart> carPartOptional = carPartRepository.findById(id);
+        if (carPartOptional.isPresent()) {
+            carPartOptional.get().getCarModels().forEach(carPart -> carPart.setCarParts(null) );
+            carPartRepository.deleteById(id);
+        } else {
+            throw new CarPartNotFoundException();
+        }
     }
 
     @Override
     public Collection<CarPart> getCarPartByName(String name) {
-        return carPartRepository.findByName(name);
+        Collection<CarPart> carParts = carPartRepository.findByName(name);
+        if (carParts.isEmpty()) {
+            throw new CarPartNotFoundException();
+        }
+        return carParts;
     }
 
     @Override
